pyproject.toml
[tool.poetry]
name = "lesson-29"
version = "0.1.0"
description = ""
authors = ["a-abdulin <abdulin@list.ru>"]
readme = "README.md"
packages = [{include = "lesson_29"}]

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
